+++ arrow-udf-python/tests/python.rs
@@ -14,9 +14,12 @@
 
 use std::sync::Arc;
 
+use arrow_array::builder::{MapBuilder, StringBuilder};
+use arrow_array::types::Int32Type;
+use arrow_array::Array;
 use arrow_array::*;
 use arrow_cast::pretty::{pretty_format_batches, pretty_format_columns};
-use arrow_schema::{DataType, Field, Schema};
+use arrow_schema::{DataType, Field, Fields, Schema};
 use arrow_udf_python::{CallMode, Runtime};
 use expect_test::{expect, Expect};
 
@@ -210,6 +213,53 @@ def json_array_access(array, i):
 }
 
 #[test]
+fn test_arg_array() {
+    let mut runtime = Runtime::new().unwrap();
+
+    runtime
+        .add_function(
+            "from_array",
+            DataType::Int32,
+            CallMode::CalledOnNullInput,
+            r#"
+def from_array(x):
+    if x:
+        return x[0]
+    else:
+        return None
+"#,
+        )
+        .unwrap();
+
+    let schema = Schema::new(vec![Field::new(
+        "x",
+        DataType::new_list(DataType::Int32, true),
+        true,
+    )]);
+    let arg0 = ListArray::from_iter_primitive::<Int32Type, _, _>([
+        Some([Some(1), Some(2)]),
+        Some([None, Some(3)]),
+        Some([Some(4), Some(5)]),
+        None,
+    ]);
+    let input = RecordBatch::try_new(Arc::new(schema), vec![Arc::new(arg0)]).unwrap();
+
+    let output = runtime.call("from_array", &input).unwrap();
+    check(
+        &[output],
+        expect![[r#"
+        +------------+
+        | from_array |
+        +------------+
+        | 1          |
+        |            |
+        | 4          |
+        |            |
+        +------------+"#]],
+    );
+}
+
+#[test]
 fn test_return_array() {
     let mut runtime = Runtime::new().unwrap();
 
@@ -247,6 +297,53 @@ def to_array(x):
 }
 
 #[test]
+fn test_arg_large_array() {
+    let mut runtime = Runtime::new().unwrap();
+
+    runtime
+        .add_function(
+            "from_large_array",
+            DataType::Int32,
+            CallMode::CalledOnNullInput,
+            r#"
+def from_large_array(x):
+    if x:
+        return x[0]
+    else:
+        return None
+"#,
+        )
+        .unwrap();
+
+    let schema = Schema::new(vec![Field::new(
+        "x",
+        DataType::new_large_list(DataType::Int32, true),
+        true,
+    )]);
+    let arg0 = LargeListArray::from_iter_primitive::<Int32Type, _, _>([
+        Some([Some(1), Some(2)]),
+        Some([None, Some(3)]),
+        Some([Some(4), Some(5)]),
+        None,
+    ]);
+    let input = RecordBatch::try_new(Arc::new(schema), vec![Arc::new(arg0)]).unwrap();
+
+    let output = runtime.call("from_large_array", &input).unwrap();
+    check(
+        &[output],
+        expect![[r#"
+        +------------------+
+        | from_large_array |
+        +------------------+
+        | 1                |
+        |                  |
+        | 4                |
+        |                  |
+        +------------------+"#]],
+    );
+}
+
+#[test]
 fn test_return_large_array() {
     let mut runtime = Runtime::new().unwrap();
 
@@ -280,6 +377,108 @@ def to_large_array(x):
         |                |
         | [3, 4, 5]      |
         +----------------+"#]],
+    );
+}
+
+#[test]
+fn test_arg_map() {
+    let mut runtime = Runtime::new().unwrap();
+
+    runtime
+        .add_function(
+            "from_map",
+            DataType::Utf8,
+            CallMode::CalledOnNullInput,
+            r#"
+def from_map(x):
+    if isinstance(x, dict) and 'k' in x:
+        return x['k']
+    else:
+        return None
+"#,
+        )
+        .unwrap();
+
+    let mut builder =
+        MapBuilder::with_capacity(None, StringBuilder::new(), StringBuilder::new(), 3);
+    builder.keys().append_value("k");
+    builder.values().append_value("v");
+    builder.append(true).unwrap();
+    builder.keys().append_value("k1");
+    builder.values().append_value("v1");
+    builder.keys().append_value("k2");
+    builder.values().append_value("v2");
+    builder.append(true).unwrap();
+    builder.append(false).unwrap();
+    let arg0 = builder.finish();
+
+    let data_type = arg0.data_type().clone();
+    let schema = Schema::new(vec![Field::new("x", data_type, true)]);
+    let input = RecordBatch::try_new(Arc::new(schema), vec![Arc::new(arg0)]).unwrap();
+
+    let output = runtime.call("from_map", &input).unwrap();
+    check(
+        &[output],
+        expect![[r#"
+        +----------+
+        | from_map |
+        +----------+
+        | v        |
+        |          |
+        |          |
+        +----------+"#]],
+    );
+}
+
+#[test]
+fn test_return_map() {
+    let mut runtime = Runtime::new().unwrap();
+
+    runtime
+        .add_function(
+            "to_map",
+            DataType::Map(
+                Arc::new(Field::new(
+                    "entries",
+                    DataType::Struct(Fields::from(vec![
+                        Field::new("keys", DataType::Utf8, true),
+                        Field::new("values", DataType::Utf8, true),
+                    ])),
+                    false,
+                )),
+                false,
+            ),
+            CallMode::CalledOnNullInput,
+            r#"
+def to_map(x, y):
+    if x is None or y is None:
+        return None
+    else:
+        return {"k1": x, "k2": y}
+"#,
+        )
+        .unwrap();
+
+    let schema = Schema::new(vec![
+        Field::new("x", DataType::Utf8, true),
+        Field::new("y", DataType::Utf8, true),
+    ]);
+    let arg0 = StringArray::from(vec![Some("ab"), None, Some("c")]);
+    let arg1 = StringArray::from(vec![Some("xy"), None, Some("z")]);
+    let input =
+        RecordBatch::try_new(Arc::new(schema), vec![Arc::new(arg0), Arc::new(arg1)]).unwrap();
+
+    let output = runtime.call("to_map", &input).unwrap();
+    check(
+        &[output],
+        expect![[r#"
+        +------------------+
+        | to_map           |
+        +------------------+
+        | {k1: ab, k2: xy} |
+        |                  |
+        | {k1: c, k2: z}   |
+        +------------------+"#]],
     );
 }